fix arithmetic error deducting stake from rejected stake - also fix logic error when calculating max rejected stake - tests
