+++ crates/consensus/worker/tests/it/main.rs
@@ -1,0 +1,7 @@
+//! IT worker tests.
+//!
+//! Use integration tests for tn-test-utils.
+mod batch_provider_tests;
+mod quorum_waiter_tests;
+
+fn main() {}
